+++ core/execution/market_snapshot.go
@@ -15,6 +15,7 @@ package execution
 import (
 	"context"
 	"fmt"
+	"sort"
 	"time"
 
 	"code.vegaprotocol.io/vega/core/assets"
@@ -33,6 +34,7 @@ import (
 	"code.vegaprotocol.io/vega/core/types"
 	"code.vegaprotocol.io/vega/libs/num"
 	"code.vegaprotocol.io/vega/logging"
+	"golang.org/x/exp/maps"
 )
 
 func NewMarketFromSnapshot(
@@ -136,7 +138,7 @@ func NewMarketFromSnapshot(
 		broker:                     broker,
 		fee:                        feeEngine,
 		liquidity:                  liqEngine,
-		parties:                    map[string]struct{}{}, // parties will be restored on PostRestore
+		parties:                    map[string]struct{}{},
 		lMonitor:                   lMonitor,
 		tsCalc:                     tsCalc,
 		feeSplitter:                NewFeeSplitterFromSnapshot(em.FeeSplitter, now),
@@ -161,6 +163,10 @@ func NewMarketFromSnapshot(
 		lpPriceRange:               mkt.LPPriceRange,
 		linearSlippageFactor:       mkt.LinearSlippageFactor,
 		quadraticSlippageFactor:    mkt.QuadraticSlippageFactor,
+	}
+
+	for _, p := range em.Parties {
+		market.parties[p] = struct{}{}
 	}
 
 	market.assetDP = uint32(assetDetails.DecimalPlaces())
@@ -190,6 +196,9 @@ func (m *Market) getState() *types.ExecMarket {
 		sp = m.settlementDataInMarket.Clone()
 	}
 
+	parties := maps.Keys(m.parties)
+	sort.Strings(parties)
+
 	em := &types.ExecMarket{
 		Market:                     m.mkt.DeepClone(),
 		PriceMonitor:               m.pMonitor.GetState(),
@@ -211,6 +220,7 @@ func (m *Market) getState() *types.ExecMarket {
 		FeeSplitter:                m.feeSplitter.GetState(),
 		SettlementData:             sp,
 		NextMTM:                    m.nextMTM.UnixNano(),
+		Parties:                    parties,
 	}
 
 	return em